GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   branch head: f442d2ff9b62a3d6a8f345004d99e021a2b35afd
   - src/components/chat/chat-input.tsx

Your previous branch was: refs/heads/main

The sha for that commit was: baebf3277907515386f93fcc32be68e94c95d521

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch